src/core/pyproject.toml
[project]
name = "adp-core"
version = "0.1.0"
description = "Agent design pattern core framework"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Ly Hon Quang", email = "lyhonquang@gmail.com" }
]
requires-python = ">=3.10.12"
dependencies = [
    "a2a-sdk>=0.3.12",
    "pandas>=2.0.0",
    "sacrebleu>=2.5.1",
]

[dependency-groups]
lint = [
    "ruff==0.14.0",
]

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
